Moved MarkerDescription to its own class
